requirements.txt
requests>=2.28.0
pandas>=1.5.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
tenacity>=8.0.0
pytest>=7.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
tiktoken>=0.5.0
tqdm>=4.65.0
aiohttp>=3.8.0
openpyxl>=3.1.0

# AI Router & RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0
